Add files via upload
It seems like you are looking for a description or information about the Titanic dataset. The Titanic dataset is a popular dataset used for machine learning and data analysis. It contains information about passengers on the Titanic, including whether they survived or not. Here is a brief description of the columns in the dataset: